initial commit
I've been working on the beginnings of Meme It So.

extract_all_thumbs.py: find all the episodes of a show and extract all
the frames as thumbnails jpg files using ffmpeg. Each filename is {frame_num}.jpg

engif: (needs to change to match extract_all_thumbs.py),
given a subtitle file, it searchs for the matching frames and creates a
gif.