react front end initial commit
Should have checked this in sooner...

Have a semi-working frontpage. Started working on LCARS design and got
carried away. So the plan for the next few commits are to clean up some
of the css mess (switching to tailwind) and then to clean up some of the
backend changes as well.