switch to using sqlite
For some reason, I really wanted to avoid using a db, and just read in
csv files and work with it in memory.

But, when you see even a few places where a join makes sense, a db
becomes too tempting.

So I've switched to loading the information into sqlite. No ORM tools,
my sql is currently integrated into the view. I can't help but say: I
kinda like this.

This really was more work than I was expecting.

Next checkin will be cleanup of the various utils that worked with
reading in the csv files.