move db initialization into flask
There's a flask way to do cli work. I've been putting cli commands for
preparation into a "prepare" directory. But instead, you create commands
that flask executes itself. This has the advantage of creating an "app
context", so the command executes from an environment similar to if it
were running from the flask server.

Anyway, as a first step, I'm moving the create_db.py into the flask
method. 'flask init-db' now creates and loads the data into the database

Also, the schema is now moved out of code and into a schema.sql, which
makes the schema easier to read.