Responsiveness
Put in some css directives (and some js too) for responsiveness.

I'm not saying this is a great layout for all devices, but it's at least
a bit sensible now in the case of mobile devices

ThumbnailNav is calculating the number of images to show based on the
viewport width. But it doesn't update on resize, so that is a bug