MemeEditor saves state when nevigating back to it
Using react router's history to save the current state of the memeditor.
Created a custome hook: useHistoryState that acts like the usual
useState, but checks the history state of the location for an initial
value, and saves changes to the location history. Is this good practice?
I don't know. Lots of people use "localState", but I that's too
permanent for this use. Maybe session state would be good?

Also switched to using base64 encoding of meme text. Using the 'urlsafe"
version of base64.